CI: deploy: ignore 'dev' directory (regression fix)
Otherwise, if the master branch is deployed, the 'dev' directory is
removed, which possibly contains data we want to keep, for example code
coverage reports or proposed website changes.

This commit fixes a regression introduced in
4e0c624ffbfa12a760acbe55ba1410bcc71749b7.

It was tested via an intermediate commit to ensure it works as expected.

Relates !71